requirements.txt
appdirs==1.4.2
beautifulsoup4==4.5.3
Django==1.10.5
djangorestframework==3.5.4
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.1.10
six==1.10.0
